created the services
I created the services for the admin, product and user. I also created the controllers for the admin, product and user. I also created the routes for the admin, product and user. I also created the middlewares for the admin, product and user. I also created the validations for the admin, product and user. I also created the utils for the admin, product and user. I also created the config for the admin, product and user. I also created the types for the admin, product and user.